further refactors and import improvements
