+++ bct/__init__.py
@@ -1,2 +1,3 @@
-from .bct import *
+from .algorithms import *
+from .utils import *
 from .nbs import *
