+++ bct/algorithms/core.py
