--- bct/bct/__init__.py
@@ -1,2 +1,0 @@
-from .algorithms import *
-from .utils import *
